♻️ extract TransportConfiguration from Configuration (#691)
